Sentimentinvestor commands
